requirements.txt
vllm
transformers
sentence-transformers
aiohttp
alfworld
beautifulsoup4
chardet
fastapi
flask
fuzzywuzzy
func-timeout
lxml
nltk
numpy
openai
openai-whisper
pandas
pdfplumber
PyPDF2
python-dateutil
python-docx
python-Levenshtein
python-pptx
pyyaml
requests
rouge
scipy
spotipy
sympy
tqdm
